requirement.txt
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-kms==2.5.0
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
crcmod==1.7
cryptography==2.3.1
Django==2.0
django-tinymce==2.7.0
idna==2.7
oss2==2.6.0
pycparser==2.19
pycryptodome==3.7.0
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
